+++ app.js
@@ -1,21 +1,24 @@
-const express = require("express");
-const path = require("path");
-const mainRoute = require("./routes/main");
-const gamesRouter = require("./routes/games");
-const bodyParser = require("body-parser");
-const cors = require("./middlewares/cors");
+const express = require('express');
+const path = require('path');
+const bodyParser = require('body-parser')
+const cors = require('./middlewares/cors');
+
+const mainRoute = require('./routes/main');
+const gamesRouter = require('./routes/games');
 
 const PORT = 3000;
+
 const app = express();
 
 app.use(
-    cors,
-    bodyParser.json(),
-    express.static(path.join(__dirname, "public")),
-    mainRoute,
-    gamesRouter
-);
+  cors,
+  bodyParser.json(),
+  express.static(path.join(__dirname, 'public')),
+  mainRoute,
+  gamesRouter
+)
+
 
 app.listen(PORT, () => {
-    console.log(`Server is running at PORT http://localhost:${PORT}`);
-});
+  console.log(`Server is running at PORT http://localhost:${PORT}`);
+})
