+++ middlewares/cors.js
@@ -1,14 +1,16 @@
 const allowedCors = [
-    "https://practicum.yandex.ru",
-    "https://students-projects.ru",
-    "localhost:3000",
+  'https://practicum.yandex.ru',
+  'https://students-projects.ru',
+  'localhost:3000'
 ];
-function cors(req, res, next) {
-    const { origin } = req.headers;
-    if (allowedCors.includes(origin)) {
-        res.header("Access-Control-Allow-Origin", origin);
-    }
 
-    next();
-}
+function cors(req, res, next) {
+  const { origin } = req.headers; 
+  if (allowedCors.includes(origin)) {
+    res.header('Access-Control-Allow-Origin', origin);
+  }
+
+  next();
+};
+
 module.exports = cors;
