+++ routes/main.js
@@ -2,8 +2,8 @@ const mainRoute = require("express").Router();
 const fs = require("fs").promises;
 
 mainRoute.get("/", (req, res) => {
-    fs.readFile("./public/index.html", "utf-8").then((data) => {
-        res.header("Content-Type", "text/html").send(data);
-    });
+  fs.readFile("./public/index.html", "utf-8").then(data => {
+    res.header("Content-Type", "text/html").send(data);
+  });
 });
 module.exports = mainRoute;
